Reorganize code to maintain import paths
