+++ stormcloud/common/cloud.py
