+++ stormcloud/common/logger.py
