+++ stormcloud/ms/meilisearch_upload.py
