+++ stormcloud/plugins/sst/entrypoint.sh
