+++ stormcloud/plugins/temp_precip/.env.sample
