+++ stormcloud/plugins/temp_precip/entrypoint.sh
