+++ stormcloud/plugins/temp_precip/main.py
