+++ stormcloud/plugins/temp_precip/temp_precip.yaml
