+++ stormcloud/storms/utils/plotter.py
